+++ tools/preferences/tool.js
@@ -45,7 +45,8 @@
         'noDirectoryPickerWarning': 'This browser does not support this functionality',
         'folderButtonTitle': 'Pick a folder',
         'fileButtonTitle': 'Pick a file',
-        'userManualLinkTitle': 'User manual (EN)'
+        'userManualLinkTitle': 'User manual (EN)',
+        'preferenceFilterPlaceholder': 'Filter preferences'
       },
       'ro-RO': {
         yes: 'Da',
@@ -76,7 +77,8 @@
         'noDirectoryPickerWarning': 'Acest browser nu suport\u0103 aceast\u0103 func\u0163ionalitate',
         'folderButtonTitle': 'Alege un director de fi\u015fiere',
         'fileButtonTitle': 'Alege un fi\u015fier',
-        'userManualLinkTitle': 'Manual utilizator (EN)'
+        'userManualLinkTitle': 'Manual utilizator (EN)',
+        'preferenceFilterPlaceholder': 'Filtru preferin\u0163e'
       }
     }
 
@@ -95,8 +97,7 @@
 
       lt.global.document.title = trans.noarg('lichessToolsPreferences');
 
-      $('nav.page-menu__menu a.active').removeClass('active');
-      $('a.lichessTools-menu').addClass('active');
+      const subnav = $('main nav.subnav').hide();
 
       const showSaved = () => {
         $('.saved').removeClass('none');
@@ -148,6 +149,9 @@
 <div class="feedback">
   <textarea enterkeyhint="send"></textarea>
   <button data-icon="${lt.icon.toEntity(lt.icon.PlayTriangle)}" title="$trans(feedbackButtonTitle)"></button>
+</div>
+<div class="prefFilter">
+  <input type="text" class="prefFilter" placeholder="$trans(preferenceFilterPlaceholder)">
 </div>`;
       }
 
@@ -168,16 +172,28 @@
         }
       }
 
-      for (const key in categs) {
+      const order = ['languages', 'community', 'general', 'appearance', 'analysis', 'analysis2', 'study', 'friends', 'play', 'puzzles', 'tv', 'mobile', 'comm', 'integration'];
+      const keys = Object.keys(categs);
+      keys.sort((a,b)=>order.indexOf(a)-order.indexOf(b));
+      for (const key of keys) {
         const categ = categs[key];
-        html += '<div><h3><label for="chk_' + key + '">$trans(options.' + key + ')</label></h3><input type="checkbox" id="chk_' + key + '" class="categoryToggle">';
+        const prefCount = categ.filter(pref=>(this.options.advanced||!pref.advanced) && !pref.hidden && (isLoggedIn || !pref.needsLogin)).length;
+        html += '<div><h3><label for="chk_' + key + '"><span class="lichessTools-prefCount">(' + prefCount + ')</span> $trans(options.' + key + ')</label></h3><input type="checkbox" id="chk_' + key + '" class="categoryToggle" checked>';
         for (const pref of categ) {
           const defaultValue = (!isLoggedIn && pref.defaultNotLoggedInValue !== undefined) ? pref.defaultNotLoggedInValue : pref.defaultValue;
 
           html += `<section data-pref="${pref.name}"`;
           const classes = [];
-          if (pref.advanced) classes.push('lichessTools-advancedPreference');
-          if (pref.hidden || (!isLoggedIn && pref.needsLogin)) classes.push('lichessTools-hiddenPreference');
+          if (pref.advanced) {
+            classes.push('lichessTools-advancedPreference');
+          }
+          if (pref.hidden || (!isLoggedIn && pref.needsLogin)) {
+            classes.push('lichessTools-hiddenPreference');
+          } else {
+            if (!pref.advanced) {
+              classes.push('lichessTools-normalPreference');
+            }
+          }
           if (pref.wip) classes.push('lichessTools-wipPreference');
           if (classes.length) html += ' class="' + classes.join(' ') + '"'
           html += `><h2>$trans(options.${pref.name})`;
@@ -274,8 +290,41 @@
       container.empty()
         .append(html)
         .addClass('lichessTools-preferences');
+
+      $('input.prefFilter',container)
+        .on('input',ev=>{
+          const tokens = (ev.target.value?.split(/\s+/) || []).filter(t=>t).map(t=>t.toLowerCase());
+          const prefs = [];
+          for (const key in categs) {
+            const categ = categs[key];
+            const items = categ
+                            .filter(pref=>(this.options.advanced||!pref.advanced) && !pref.hidden && (isLoggedIn || !pref.needsLogin))
+                            .map(pref=>({ 
+                               name: pref.name,
+                               text: [pref.name, pref.category, trans.noarg('options.'+pref.name),trans.noarg('options.'+pref.category)]
+                                       .concat((pref.possibleValues||[]).map(v=>{
+                                         if (typeof(v) === 'boolean') return v?'yes':'no';
+                                         return pref.name+'.'+v;
+                                       }).map(v=>trans.noarg(v)))
+                                       .concat((pref.possibleValues||[]).map(v=>{
+                                         if (typeof(v) !== 'boolean') return pref.name+'.'+v;
+                                       }))
+                                       .join(' ').toLowerCase()
+                             }));
+            prefs.push.apply(prefs,items);
+          }
+          const visiblePrefs = prefs.filter(p=>!tokens.find(t=>!p.text.includes(t))).map(p=>p.name);
+          $('section[data-pref]').each((i,e)=>{
+            const pref = $(e).attr('data-pref');
+            const shown = visiblePrefs.includes(pref);
+            $(e)
+              .toggleClass('filteredIn',shown && tokens.length)
+              .toggleClass('filteredOut',!shown);
+          });
+        });
+
       $('form', container).append(saved);
-      $('form input:not(.categoryToggle)', container)
+      $('form input:not(.categoryToggle,.prefFilter)', container).add($('main nav.subnav input:not(.categoryToggle,.prefFilter)'))
         .each((i, e) => {
           const type = $(e).prop('type');
           const isCheckable = type == 'radio' || type == 'checkbox';
@@ -539,7 +588,7 @@
           .attr('data-icon', lt.icon.InfoCircle)
           .attr('href', 'https://siderite.dev/blog/lichess-tools---user-manual/#' + pref)
           .attr('target', '_blank')
-          .prependTo(e);
+          .appendTo($('h2',e));
       });
     };
 
@@ -548,17 +597,20 @@
       const trans = lt.translator;
       const $ = lt.$;
       if (!$('#dasher_app a.lichessTools-preferences').length) {
+        const elem = $('<a class="text lichessTools-preferences">')
+          .attr('data-icon', lt.icon.Gear)
+          .attr('href', '/team/all#lichessTools')
+          .text(trans.noarg('lichessTools'))
+          .attr('title', trans.noarg('lichessToolsPreferences'));
         let links = $('#dasher_app div.links');
         if (!links.length) {
           links = $('<div class="links">')
             .insertBefore('#dasher_app div.subs');
+          elem.appendTo(links);
+        } else {
+          elem
+            .insertAfter($('[href="/account/profile"]',links));
         }
-        $('<a class="text lichessTools-preferences">')
-          .attr('data-icon', lt.icon.Gear)
-          .attr('href', '/team/all#lichessTools')
-          .text(trans.noarg('lichessTools'))
-          .attr('title', trans.noarg('lichessToolsPreferences'))
-          .appendTo(links);
       }
     };
 
@@ -574,17 +626,16 @@
       };
       this.logOption('Integration in Preferences', this.options.enabled);
       this.logOption(' ... show advanced', this.options.advanced);
+      if (!this.options.enabled) return;
 
-      const isLoggedIn = !!lt.getUserId();
-      const isLoggedOutTeams = location.pathname == '/team/all' && !isLoggedIn;
-      lt.global.clearInterval(this.interval);
-      if (!isLoggedIn) {
-        this.interval = lt.global.setInterval(this.addPreferencesMenu,500);
-      }
-
-      if (!$('main.account').length && !isLoggedOutTeams) {
+      if (/^\/account\/preferences\/display#lichessTools/.test(lt.global.location.pathname+lt.global.location.hash)) {
+        lt.global.location = '/team/all'+lt.global.location.hash;
         return;
       }
+
+      const isTeams = location.pathname == '/team/all';
+      lt.global.clearInterval(this.interval);
+      this.interval = lt.global.setInterval(this.addPreferencesMenu,500);
 
       if ($('a.lichessTools-menu').length) return;
       const openPreferences = this.openPreferences;
@@ -592,21 +643,23 @@
       let $this = this;
       const f = function () {
         if (location.hash?.startsWith('#lichessTools')) {
-          if (isLoggedOutTeams) {
-            lichess.asset.loadCssPath('bits.account');
-            $('main nav.subnav').hide();
-          }
+          lichess.asset.loadCssPath('bits.account');
           openPreferences();
-          lt.global.setTimeout(() => {
+          const fc = () => {
             const m = /#lichessTools\/(?<pref>.*)$/.exec(location.hash);
             const pref = m?.groups?.pref;
-            if (pref != $this._lastScrolled) {
+            if (pref && pref != $this._lastScrolled) {
               const elem = $('[data-pref="' + pref + '"]');
+              if (!elem.length) {
+                lt.global.setTimeout(fc, 500);
+                return;
+              }
               elem.parents().add(elem).show();
               elem[0]?.scrollIntoView();
               $this._lastScrolled = pref;
             }
-          }, 500);
+          };
+          fc();
         } else {
           if ($('.lichessTools-preferences').length) {
             location.reload();
@@ -615,12 +668,6 @@
       };
       $(lt.global).on('hashchange', f);
 
-      const prefElem = $('<a>')
-        .addClass('lichessTools-menu')
-        .attr('href', isLoggedOutTeams ? '/team/all#lichessTools' : '/account/preferences/display#lichessTools')
-        .append($('<span>').text(trans.noarg('LiChess Tools')));
-
-      $('.page-menu__menu a[href*=privacy]').before(prefElem);
       f();
     }