+++ tools/puzzleOptions/tool.js
@@ -8,9 +8,8 @@
         name: 'puzzleOptions',
         category: 'puzzles',
         type: 'multiple',
-        possibleValues: ['wakeLock','endTimer', 'showSessionTotal'],
-        defaultValue: 'wakeLock,showSessionTotal',
-        advanced: true
+        possibleValues: ['endTimer', 'showSessionTotal'],
+        defaultValue: 'showSessionTotal'
       }
     ];
 
@@ -18,7 +17,6 @@
       'en-US': {
         'options.puzzles': 'Puzzles',
         'options.puzzleOptions': 'Options for puzzles',
-        'puzzleOptions.wakeLock': 'Prevent screen lock playing puzzles',
         'puzzleOptions.endTimer': 'Show completion time',
         'puzzleOptions.showSessionTotal': 'Show session total',
         'elapsedText': '(%s s)',
@@ -27,7 +25,6 @@
       'ro-RO': {
         'options.puzzles': 'Probleme de \u015Fah',
         'options.puzzleOptions': 'Op\u0163iuni pentru probleme de \u015fah',
-        'puzzleOptions.wakeLock': 'Previne blocarea ecranului \u00een probleme de \u015fah',
         'puzzleOptions.endTimer': 'Arat\u0103 durata complet\u0103rii',
         'puzzleOptions.showSessionTotal': 'Arat\u0103 totalul per sesiune',
         'elapsedText': '(%s s)',
@@ -44,34 +41,6 @@
     isTrainingPage = ()=>{
       const lt = this.lichessTools;
       return /^\/training/i.test(lt.global.location.pathname) && !/^\/training\/(?:dashboard|themes)/.test(lt.global.location.pathname);
-    };
-
-    handleWakeLock = async () => {
-      const lt = this.lichessTools;
-      if (this.isPlaying()) {
-        if (!this.wakelock) {
-          await this.requestWakeLock();
-        }
-      } else {
-        lt.global.clearTimeout(this.wakeLockTimeout);
-        this.wakelock?.release();
-        this.wakelock=null;
-      }
-    };
-
-    requestWakeLock = async () => {
-      const lt = this.lichessTools;
-      try {
-        if (document.visibilityState === 'visible') {
-          this.wakelock?.release();
-          this.wakelock = await lt.global.navigator.wakeLock.request("screen");
-          if (this.wakelock) return;
-        }
-      } catch (err) {
-        console.debug('Wakelock failed:', err);
-      }
-      lt.global.clearTimeout(this.wakeLockTimeout);
-      this.wakeLockTimeout = lt.global.setTimeout(this.requestWakeLock, 1000);
     };
 
     showTotal = () => {
@@ -125,21 +94,9 @@
       const value = lt.currentOptions.getValue('puzzleOptions');
       this.logOption('Puzzle options', value);
       this.options = {
-        wakeLock: lt.isOptionSet(value, 'wakeLock'),
         endTimer: lt.isOptionSet(value, 'endTimer'),
         showSessionTotal: lt.isOptionSet(value, 'showSessionTotal')
       };
-      lt.pubsub.off('lichessTools.redraw',this.handleWakeLock);
-      lt.pubsub.off('lichessTools.puzzleStart',this.handleWakeLock);
-      if (this.options.wakeLock && this.isTrainingPage()) {
-        lt.pubsub.on('lichessTools.redraw',this.handleWakeLock);
-        lt.pubsub.on('lichessTools.puzzleStart',this.handleWakeLock);
-        await this.handleWakeLock();
-      } else {
-        lt.global.clearTimeout(this.wakeLockTimeout);
-        this.wakelock?.release();
-        this.wakelock=null;
-      }
       lt.pubsub.off('lichessTools.puzzleStart',this.startTimer);
       lt.pubsub.off('lichessTools.puzzleEnd',this.endTimer);
       if (this.options.endTimer && this.isTrainingPage()) {